requirements.txt
flask
rembg
pillow  